+++ src-tauri/src/db/schema/create.rs
@@ -1,0 +1,157 @@
+use super::{DatabaseConnection, SCHEMA_VERSION};
+use crate::crypto::cipher;
+use aes_gcm::aead::rand_core::RngCore;
+use rusqlite::Connection;
+use zeroize::Zeroize;
+
+/// Creates a new encrypted diary database (schema v5)
+///
+/// Generates a random master key, wraps it with the password, and stores the
+/// wrapped key in `auth_slots`. Entries are encrypted with the master key.
+pub fn create_database<P: AsRef<std::path::Path>>(
+    db_path: P,
+    password: String,
+) -> Result<DatabaseConnection, String> {
+    let mut master_key_bytes = [0u8; 32];
+    aes_gcm::aead::OsRng.fill_bytes(&mut master_key_bytes);
+    let encryption_key =
+        cipher::Key::from_slice(&master_key_bytes).ok_or("Invalid master key size")?;
+
+    let conn =
+        Connection::open(&db_path).map_err(|e| format!("Failed to create database: {}", e))?;
+
+    create_schema(&conn)?;
+
+    let method = crate::auth::password::PasswordMethod::new(password);
+    let wrapped_key = method
+        .wrap_master_key(&master_key_bytes)
+        .map_err(|e| format!("Failed to wrap master key: {}", e))?;
+
+    let now = chrono::Utc::now().to_rfc3339();
+    conn.execute(
+        "INSERT INTO auth_slots (type, label, wrapped_key, created_at) VALUES ('password', 'Password', ?1, ?2)",
+        rusqlite::params![&wrapped_key, &now],
+    )
+    .map_err(|e| format!("Failed to create password slot: {}", e))?;
+
+    master_key_bytes.zeroize();
+
+    Ok(DatabaseConnection {
+        conn,
+        encryption_key,
+    })
+}
+
+/// Creates a new encrypted diary database using a device-generated auto key.
+///
+/// The auto_key must be 32 random bytes generated by the caller.
+/// It is stored externally (in config.json) and never inside the database.
+pub fn create_database_auto<P: AsRef<std::path::Path>>(
+    db_path: P,
+    auto_key_bytes: &[u8; 32],
+) -> Result<DatabaseConnection, String> {
+    let mut master_key_bytes = [0u8; 32];
+    aes_gcm::aead::OsRng.fill_bytes(&mut master_key_bytes);
+    let encryption_key =
+        cipher::Key::from_slice(&master_key_bytes).ok_or("Invalid master key size")?;
+
+    let conn = Connection::open(db_path.as_ref())
+        .map_err(|e| format!("Failed to create database: {}", e))?;
+
+    create_schema(&conn)?;
+
+    let method = crate::auth::auto_key::AutoKeyMethod { auto_key_bytes };
+    let wrapped_key = method
+        .wrap_master_key(&master_key_bytes)
+        .map_err(|e| format!("Failed to wrap master key: {}", e))?;
+
+    let now = chrono::Utc::now().to_rfc3339();
+    conn.execute(
+        "INSERT INTO auth_slots (type, label, wrapped_key, created_at) VALUES ('auto', 'Local Key', ?1, ?2)",
+        rusqlite::params![&wrapped_key, &now],
+    )
+    .map_err(|e| format!("Failed to create auto slot: {}", e))?;
+
+    master_key_bytes.zeroize();
+
+    Ok(DatabaseConnection {
+        conn,
+        encryption_key,
+    })
+}
+
+fn create_schema(conn: &Connection) -> Result<(), String> {
+    conn.execute_batch(
+        r#"
+        -- Schema version table
+        CREATE TABLE IF NOT EXISTS schema_version (
+            version INTEGER PRIMARY KEY
+        );
+
+        -- Metadata table (kept for backward compatibility with v1/v2 migrations)
+        CREATE TABLE IF NOT EXISTS metadata (
+            key TEXT PRIMARY KEY,
+            value TEXT NOT NULL
+        );
+
+        -- Entries table (encrypted data, multiple entries per day supported via AUTOINCREMENT id)
+        CREATE TABLE IF NOT EXISTS entries (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            date TEXT NOT NULL,
+            title_encrypted BLOB,
+            text_encrypted BLOB,
+            word_count INTEGER DEFAULT 0,
+            date_created TEXT NOT NULL,
+            date_updated TEXT NOT NULL
+        );
+        CREATE INDEX IF NOT EXISTS idx_entries_date ON entries(date);
+
+        -- Search index: not implemented. A future search module should create its index here.
+        -- Interface contract: see commands/search.rs for SearchResult and the search_entries command.
+
+        -- Authentication slots (password, keypair, etc.)
+        CREATE TABLE IF NOT EXISTS auth_slots (
+            id          INTEGER PRIMARY KEY AUTOINCREMENT,
+            type        TEXT NOT NULL,
+            label       TEXT NOT NULL,
+            public_key  BLOB,
+            wrapped_key BLOB NOT NULL,
+            created_at  TEXT NOT NULL,
+            last_used   TEXT
+        );
+
+        -- Key-value store for journal-level settings (require_all_auth, etc.)
+        CREATE TABLE IF NOT EXISTS db_settings (
+            key   TEXT PRIMARY KEY,
+            value TEXT NOT NULL
+        );
+
+        -- Tags (encrypted name + HKDF-keyed fingerprint for dedup)
+        CREATE TABLE IF NOT EXISTS tags (
+            id                INTEGER PRIMARY KEY AUTOINCREMENT,
+            name_encrypted    BLOB    NOT NULL,
+            name_fingerprint  TEXT    NOT NULL UNIQUE,
+            created_at        TEXT    NOT NULL
+        );
+
+        -- Entry-tag associations
+        CREATE TABLE IF NOT EXISTS entry_tags (
+            entry_id  INTEGER NOT NULL,
+            tag_id    INTEGER NOT NULL,
+            PRIMARY KEY (entry_id, tag_id),
+            FOREIGN KEY (entry_id) REFERENCES entries(id) ON DELETE CASCADE,
+            FOREIGN KEY (tag_id)   REFERENCES tags(id)    ON DELETE CASCADE
+        );
+        CREATE INDEX IF NOT EXISTS idx_entry_tags_tag_id ON entry_tags(tag_id);
+        "#,
+    )
+    .map_err(|e| format!("Failed to create schema: {}", e))?;
+
+    conn.execute(
+        "INSERT OR REPLACE INTO schema_version (version) VALUES (?1)",
+        [SCHEMA_VERSION],
+    )
+    .map_err(|e| format!("Failed to set schema version: {}", e))?;
+
+    Ok(())
+}
